# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "микросервис по сотрудникам",
        "title": "employees-manager",
        "contact": {
            "name": "Pavel Utkin",
            "url": "https://t.me/utkin_pawka",
            "email": "pavel@utkin-pro.ru"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/employee/find{searchString}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "GET"
                ],
                "summary": "Найти сотрудника по ФИО.",
                "operationId": "findGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ФИО",
                        "name": "searchString",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/employee/get-number-of-vacation-days{employeeID}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "GET"
                ],
                "summary": "Получить количество дней отпуска.",
                "operationId": "getNumberOfVacationDaysGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "идентификатор сотрудника",
                        "name": "employeeID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/employee/hire": {
            "post": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "Нанять сотрудника.",
                "operationId": "hirePOST",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ФИО",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Пол",
                        "name": "gender",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Возраст",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Телефон",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Электронная почта",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Адрес прописки",
                        "name": "address",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "description": "\"микросервис по сотрудникам\"",
            "name": "Update"
        },
        {
            "description": "\"микросервис по сотрудникам\"",
            "name": "Value"
        },
        {
            "description": "\"микросервис по сотрудникам\"",
            "name": "Static"
        }
    ]
}